requirements.txt
# === Core ===
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
gunicorn==22.0.0

# === Database ===
psycopg2-binary==2.9.9       # PostgreSQL (Supabase free tier)
dj-database-url==2.2.0

# === AI Models (all have free tiers) ===
google-generativeai==0.7.2   # Gemini 1.5 Flash — very generous free tier
groq==0.9.0                  # Groq — free tier
httpx==0.27.0                # OpenRouter — HTTP calls (free models available)

# === MCP Protocol ===
mcp==1.0.0                   # Official MCP Python SDK

# === Document Processing ===
pypdf==4.3.1                 # PDF text extraction
python-docx==1.1.2           # DOCX extraction
pillow==10.4.0               # Image handling

# === Task Queue (optional but recommended) ===
celery==5.4.0
redis==5.0.8                 # Upstash Redis free tier

# === Utilities ===
pydantic==2.8.2              # Validation
python-multipart==0.0.9      # File uploads
structlog==24.4.0            # Structured logging
tenacity==8.5.0              # Retry logic
aiohttp==3.10.5              # Async HTTP
whitenoise==6.7.0            # Static files (no S3 needed)
